Removed cookie banner.

+++ cleanup.php
@@ -5,7 +5,7 @@ foreach($databases as $db)
     $file = "db/$db";
     $olderThanOneday = (time() - filemtime($file)) > 86400;
     // Exclude hidden files and symlinks.
-    if (substr($db, 0, 1) != "." && $olderThanOneday)
+    if (substr($db, 0, 1) != "." && $olderThanOneday && substr($db, -3, 3) === '.db')
     {
         unlink($file);
     }
